+++ rubato/utils/__init__.py
@@ -4,7 +4,7 @@ This module houses low-dependency utilities widely used in rubato.
 from .path import get_path
 from .error import *
 from .rb_time import DelayedTask, FramesTask, RecurrentTask, Time
-from .math import *
+from .computation import *
 from .hardware import *
 from .radio import *
 from .color import Color
